Formalization Added
I just typed up a more detailed formalization of a possible approach to evaluating the individual characters based on an approach which considers them "region by region" (more detailed description in the PDF). I then talk a bit about how to interpret the quantification of the accuracy to the printed image we get from our CNN and incorporate those quantities into a big picture analysis that's useful for the user.