added some straightforward functions to cld_lookup.py (still useful!) added a new file, decomp_lookup.py, which as of now will get you components the way we want, and we can use that in conjunction with cld_lookup. I think the best way to break things up as of now is just to interpret the character string after : and segment accordingly by matching (this is intentionally vague because I don't want to go down the rabbit hole right now). Long story short, it's a very powerful decomposition database, but we need to be careful to only use as much power as we need (it will break things down into strokes if you're not careful).
The most important thing now is to keep things simple. Have the basic
segmentation and work with that. We can recursively break up the character further from
there, but baby steps for now. We need to be able to effectively segment
the basic patterns so that the recursive breaking (i.e. just doing the
same thing inductively) works almost trivially. If we can get up-down,
left-right, inside-outside,... working fine, then the recursion is, as I
said, effectively trivial. So anyways, I think we've got a good start on
segmenting, but we gotta move quickly.